// src/common/interfaces/WellboreData.ts
export interface WellboreData {
  well_id: string;
  wellbore_id: string;
  tight_group_id: string;
  ko_date: string;
  authorized_md: number;
  phase: string;
  authorized_tvd: number;
  bh_md: number;
  bh_tvd: number;
  is_deviated: string;
  budgeted_md: number;
  geo_offset_east_bh: number;
  budgeted_tvd: number;
  geo_offset_north_bh: number;
  geo_latitude_bh: number;
  geo_longitude_bh: number;
  geo_offset_east_ko: number;
  reason: string;
  geo_offset_north_ko: number;
  geo_latitude_ko: number;
  ko_md: number;
  ko_tvd: number;
  geo_longitude_ko: number;
  end_status: string;
  well_legal_name: string;
  plugback_md: number;
  wellbore_name: string;
  plugback_tvd: number;
  is_readonly: string;
  wellbore_no: string;
  create_date: string;
  create_user_id: string;
  create_app_id: string;
  update_date: string;
  update_user_id: string;
  update_app_id: string;
  ow_well_uwi: string;
  external_well_id: string;
  api_no: string;
  wellbore_uwi: string;
}